Refactor, split each asset report generator to separate file
