+++ features/il_reports/lib/il_reports.dart
@@ -1,3 +1,5 @@
 library il_reports;
 
-export 'src/ui/asset_report_generators.dart';
+export 'src/ui/generators/asset_heatmap_report_generator.dart';
+export 'src/ui/generators/asset_tailmap_report_generator.dart';
+export 'src/ui/generators/asset_zone_retention_report_generator.dart';
